requirements.txt
docopt
scipy
numpy==1.16.0
pandas
biopandas
progressbar2
scikit-learn
dpu-utils>=0.1.30
tensorflow-gpu==1.13.1